+++ rollup.config.js
@@ -1,7 +1,7 @@
-import nodeResolve from 'rollup-plugin-node-resolve';
-import { uglify } from 'rollup-plugin-uglify';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
+import { terser } from 'rollup-plugin-terser';
 import sourcemaps from 'rollup-plugin-sourcemaps';
-import commonjs from 'rollup-plugin-commonjs';
+import commonjs from '@rollup/plugin-commonjs';
 
 let MINIFY = process.env.MINIFY;
 
@@ -13,13 +13,20 @@ let banner = `/**
  * @license MIT License, http://www.opensource.org/licenses/MIT
  */`;
 
-let uglifyOpts = { output: {} };
-// retain multiline comment with @license
-uglifyOpts.output.comments = (node, comment) => comment.type === 'comment2' && /@license/i.test(comment.value);
+let plugins = [nodeResolve(), sourcemaps(), commonjs()];
 
-let plugins = [nodeResolve({ jsnext: true }), sourcemaps(), commonjs()];
-
-if (MINIFY) plugins.push(uglify(uglifyOpts));
+if (MINIFY) {
+  plugins.push(
+    terser({
+      format: {
+        // retain multiline comment with @license
+        comments: (node, comment) => {
+          return comment.type == 'comment2' && /@license/i.test(comment.value);
+        },
+      },
+    })
+  );
+}
 
 let extension = MINIFY ? '.min.js' : '.js';
 
@@ -42,14 +49,14 @@ function isExternal(id) {
   ];
 
   let regexps = externals
-    .map(e => [
+    .map((e) => [
       new RegExp(`^${e}$`),
       // new RegExp(`commonjs-proxy.${e}$`),
       new RegExp(`node_modules/${e}`),
     ])
     .reduce((acc, a) => acc.concat(a), []);
 
-  return regexps.map(regex => regex.exec(id)).reduce((acc, val) => acc || !!val, false);
+  return regexps.map((regex) => regex.exec(id)).reduce((acc, val) => acc || !!val, false);
 }
 
 const CONFIG = {