+++ src/react/UIRouterReactContext.tsx
@@ -62,11 +62,12 @@ export function UIRouterContextComponent(props: {
 
   // We know the AngularJS state. Now render the {children}
   const childrenCount = React.Children.count(children);
+  const isArray = Array.isArray(children);
 
   return (
     <UIRouterContext.Provider value={(inherited && routerFromReactContext) || contextFromAngularJS.router}>
       <UIViewContext.Provider value={(inherited && parentUIViewFromReactContext) || contextFromAngularJS.addr}>
-        {childrenCount === 1 ? React.Children.only(children) : <div>{children}</div>}
+        {childrenCount === 1 && !isArray ? React.Children.only(children) : <div>{children}</div>}
       </UIViewContext.Provider>
     </UIRouterContext.Provider>
   );