+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## 1.0.5 (2022-02-28)
+[Compare `@uirouter/react-hybrid` versions 1.0.4 and 1.0.5](https://github.com/ui-router/react-hybrid/compare/1.0.4...1.0.5)
+
+
+---
+
+
+### Updated `@uirouter/react` from 1.0.6 to 1.0.7
+
+
+Changelog for `@uirouter/react`:
+
+
+[Compare `@uirouter/react` versions 1.0.6 and 1.0.7](https://github.com/ui-router/react/compare/1.0.6...1.0.7)
+
+### Bug Fixes
+
+* update classnames library which has default export type ([#939](https://github.com/ui-router/react/issues/939)) ([c771405](https://github.com/ui-router/react/commit/c771405))
+
+## 1.0.4 (2021-05-02)
+[Compare `@uirouter/react-hybrid` versions 1.0.3 and 1.0.4](https://github.com/ui-router/react-hybrid/compare/1.0.3...1.0.4)
+
+### Bug Fixes
+
+* **UIRouterContextComponent:** Handle array of length 1 ([#526](https://github.com/ui-router/react-hybrid/issues/526)) ([925f816](https://github.com/ui-router/react-hybrid/commit/925f816))
+
 ## 1.0.3 (2020-12-23)
 [Compare `@uirouter/react-hybrid` versions 1.0.2 and 1.0.3](https://github.com/ui-router/react-hybrid/compare/1.0.2...1.0.3)